Change getCount on Clock from double to long

+++ odin-core/src/main/java/com/purplepip/odin/sequence/Clock.java
@@ -26,5 +26,5 @@ public interface Clock extends MicrosecondPositionProvider {
   /*
    * Get the current tock count.
    */
-  double getCount();
+  long getCount();
 }
